[FIX] stock: hide reset to draft button when no needed
Only immediate transfers should have the ability to reset the state to
draft. Also rewording it as "Plan" as the goal of this button is to left
the stock move to be done later and impact the forecast

Task: 3256447
Part-of: odoo/odoo#124409

+++ addons/stock/models/stock_picking.py
@@ -932,9 +932,12 @@ class Picking(models.Model):
         picking_to_reset = self.filtered(lambda p: p.state not in ('done', 'cancel'))
         picking_to_reset.do_unreserve()
         picking_to_reset.immediate_transfer = False
-        picking_to_reset.move_ids.quantity_done = 0
-        picking_to_reset.move_ids.move_line_ids.unlink()
-        picking_to_reset.move_ids.state = 'draft'
+        if picking_to_reset.move_ids:
+            picking_to_reset.move_ids.quantity_done = 0
+            picking_to_reset.move_ids.move_line_ids.unlink()
+            picking_to_reset.move_ids.state = 'draft'
+        else:
+            picking_to_reset.state = 'draft'
 
     def _action_done(self):
         """Call `_action_done` on the `stock.move` of the `stock.picking` in `self`.
